requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
